move headers to include, add headers for bodyprog and dynamic
