+++ include/bodyprog/bodyprog.h
@@ -1,0 +1,14 @@
+#ifndef _BODYPROG_H
+#define _BODYPROG_H
+
+#include "common.h"
+
+/** Declarations for unknown symbols in bodyprog. */
+
+/** Unknown bodyprog var. Set in `fsQueueDoThingWhenEmpty`. */
+extern s32 D_800C489C;
+
+/** Unknown bodyprog func. Called by `fsQueueDoThingWhenEmpty`. */
+s32 func_8003c850();
+
+#endif
