+++ src/main/mem.c
@@ -1,4 +1,4 @@
-#include "mem.h"
+#include "main/mem.h"
 
 void memmove32(u32 *dest, u32 *src, s32 numbytes) {
   u32 *end;
